requirements.txt
aiogram==3.21.0
asyncio==4.0.0
aiohttp==3.12.15
python-dotenv==1.0.1